update positions for pop-up windows
