+++ src/plugins/frameSettings/FrameSettingsEditor.tsx
@@ -1,5 +1,4 @@
 import React, { HTMLAttributes, useEffect, useMemo, useRef, useState } from 'react';
-import { usePopper } from 'react-popper';
 import { IconButton, Portal, Stack, useStyles2 } from '@grafana/ui';
 import { DataFrame, FieldType, GrafanaTheme2 } from '@grafana/data';
 import useClickAway from 'react-use/lib/useClickAway';
@@ -63,9 +62,6 @@ export const FrameSettingsEditor: React.FC<FrameSettingsEditorProps> = ({
 }) => {
   const styles = useStyles2(getStyles);
 
-  const [popperTrigger, setPopperTrigger] = useState<HTMLDivElement | null>(null);
-  const [editorPopover, setEditorPopover] = useState<HTMLDivElement | null>(null);
-
   const [size, setSize] = useState({
     width: 0,
     height: 0,
@@ -77,6 +73,7 @@ export const FrameSettingsEditor: React.FC<FrameSettingsEditorProps> = ({
   const [editableTableData, setEditableTableData] = useState<FieldSettingItem[]>([]);
 
   const clickAwayRef = useRef<HTMLDivElement>(null);
+  const wrapEditorRef = useRef<HTMLDivElement>(null);
 
   const fields = useMemo(
     () =>
@@ -103,19 +100,6 @@ export const FrameSettingsEditor: React.FC<FrameSettingsEditorProps> = ({
 
   useClickAway(clickAwayRef, () => {
     onDismiss();
-  });
-
-  const popper = usePopper(popperTrigger, editorPopover, {
-    modifiers: [
-      { name: 'arrow', enabled: false },
-      {
-        name: 'preventOverflow',
-        enabled: true,
-        options: {
-          rootBoundary: 'viewport',
-        },
-      },
-    ],
   });
 
   /**
@@ -180,12 +164,55 @@ export const FrameSettingsEditor: React.FC<FrameSettingsEditorProps> = ({
     ];
   }, [fieldSettings]);
 
+  useEffect(() => {
+    const el = wrapEditorRef.current;
+    if (!el) {
+      return;
+    }
+
+    let frameId: number;
+
+    const checkPosition = () => {
+      const rect = el.getBoundingClientRect();
+      /**
+       * Checks
+       */
+      const isPanelOutOfBottomView = rect.bottom > window.innerHeight;
+      const isPanelOutOfRightView = rect.right > window.innerWidth;
+
+      if (isPanelOutOfBottomView) {
+        const offset = rect.bottom - window.innerHeight;
+        /**
+         * 50px gap near bottom
+         */
+        const updatedTop = rect.top - offset - 50;
+        el.style.top = `${updatedTop}px`;
+      }
+
+      if (isPanelOutOfRightView) {
+        const offsetRight = rect.right - window.innerWidth;
+        /**
+         * 50px gap near bottom
+         */
+        const updatedLeft = rect.left - offsetRight - 50;
+        el.style.left = `${updatedLeft}px`;
+      }
+
+      frameId = requestAnimationFrame(checkPosition);
+    };
+
+    checkPosition();
+
+    return () => {
+      cancelAnimationFrame(frameId);
+    };
+  }, []);
+
   return (
     <Portal>
       <>
-        <div ref={setPopperTrigger} style={style} />
-        <div ref={setEditorPopover} style={popper.styles.popper} {...popper.attributes.popper}>
-          <div className={styles.editor} ref={clickAwayRef}>
+        <div ref={wrapEditorRef} style={style}>
+          <div className={styles.editor} ref={clickAwayRef} style={{ border: '1px solid blue' }}>
             <div className={styles.header}>
               <Stack justifyContent={'space-between'} alignItems={'center'}>
                 <div className={styles.title}>Set Frame settings</div>
